+++ tests/test_end_to_end/fastapi/__init__.py
@@ -1,8 +1,10 @@
 """ A FastAPI app used to create an OpenAPI document for end-to-end testing """
 import json
+from enum import Enum
 from pathlib import Path
+from typing import List
 
-from fastapi import FastAPI
+from fastapi import APIRouter, FastAPI, Query
 from pydantic import BaseModel
 
 app = FastAPI(title="My Test API", description="An API for testing openapi-python-client",)
@@ -17,6 +19,39 @@ async def ping():
     """ A quick check to see if the system is running """
     return {"success": True}
 
+
+test_router = APIRouter()
+
+
+class TestEnum(Enum):
+    """ For testing Enums in all the ways they can be used """
+
+    FIRST_VALUE = "FIRST_VALUE"
+    SECOND_VALUE = "SECOND_VALUE"
+
+
+class OtherModel(BaseModel):
+    """ A different model for calling from TestModel """
+
+    a_value: str
+
+
+class TestModel(BaseModel):
+    """ A Model for testing all the ways custom objects can be used """
+
+    an_enum_value: TestEnum
+    a_list_of_enums: List[TestEnum]
+    a_list_of_strings: List[str]
+    a_list_of_objects: List[OtherModel]
+
+
+@test_router.get("/", response_model=List[TestModel])
+def test_getting_lists(statuses: List[TestEnum] = Query(...),):
+    """ Get users, filtered by statuses """
+    return
+
+
+app.include_router(test_router, prefix="/tests", tags=["users"])
 
 if __name__ == "__main__":
     path = Path(__file__).parent / "openapi.json"