+++ tests/test_end_to_end/golden-master/my_test_api_client/api/default.py
@@ -4,8 +4,8 @@ from typing import Dict, List, Optional, Union
 import httpx
 
 from ..client import AuthenticatedClient, Client
+from ..errors import ApiResponseError
 from ..models.abc_response import ABCResponse
-from .errors import ApiResponseError
 
 
 def ping_ping_get(