+++ tests/test_end_to_end/golden-master/my_test_api_client/models/abc_response.py
@@ -20,4 +20,5 @@ class ABCResponse:
     def from_dict(d: Dict) -> ABCResponse:
 
         success = d["success"]
+
         return ABCResponse(success=success,)