+++ tests/test_end_to_end/golden-master/my_test_api_client/models/h_t_t_p_validation_error.py
@@ -1,0 +1,27 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from datetime import datetime
+from typing import Dict, List, Optional, cast
+
+from .validation_error import ValidationError
+
+
+@dataclass
+class HTTPValidationError:
+    """  """
+
+    detail: Optional[List[ValidationError]] = None
+
+    def to_dict(self) -> Dict:
+        return {
+            "detail": self.detail if self.detail is not None else None,
+        }
+
+    @staticmethod
+    def from_dict(d: Dict) -> HTTPValidationError:
+
+        detail = []
+        for detail_item in d.get("detail", []):
+            detail.append(ValidationError.from_dict(detail_item))
+        return HTTPValidationError(detail=detail,)