+++ tests/test_end_to_end/golden-master/my_test_api_client/models/statuses.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Statuses(str, Enum):
+    FIRST_VALUE = "FIRST_VALUE"
+    SECOND_VALUE = "SECOND_VALUE"