+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 - 2020-03-22
+### Fixes
+- Fixed import of errors.py in generated api modules
+
+### Additions
+- Support for lists of Enums
+
 ## 0.2.0 - 2020-03-22
 ### Changes
 - Update Typer dependency to 0.1.0 and remove click-completion dependency (#19)